Change Id type in UserActivity from uint to long for better compatibility

+++ backend/src/Core/UserActivity.cs
@@ -2,7 +2,7 @@
 
 public class UserActivity
 {
-    public uint Id { get; init; }
+    public long Id { get; init; }
     public required Guid UserId { get; init; }
     public User? User { get; init; }
     public required string Code { get; init; }
